Move tests into tests/ folder and split responses to .json files
