+++ utils/tests/data/package_mismatch_response.json
@@ -1,0 +1,19 @@
+{
+  "requestDetails": {
+    "requestPackageName": "my.own.package",
+    "timestampMillis": "1617893780",
+    "requestHash": "aGVsbG8gd29scmQgdGhlcmU"
+  },
+  "appIntegrity": {
+    "appRecognitionVerdict": "PLAY_RECOGNIZED",
+    "packageName": "my.own.package",
+    "certificateSha256Digest": ["6a6a1474b5cbbb2b1aa57e0bc3"],
+    "versionCode": "42"
+  },
+  "deviceIntegrity": {
+    "deviceRecognitionVerdict": ["MEETS_DEVICE_INTEGRITY"]
+  },
+  "accountDetails": {
+    "appLicensingVerdict": "LICENSED"
+  }
+}
